Re-organizing some more
