+++ browser/verifier-store-oidc/src/utils/cloneIndex.ts
@@ -1,0 +1,6 @@
+import { Item } from "../atoms/atoms";
+
+export const cloneIndex = (items: Item[], id: string) => ({
+  clone: items.map((item) => ({ ...item })),
+  index: items.findIndex((item) => item.id === id),
+});
